Changes to delete/modify dialogs
Adds: opening a delete/modify dialog now properly shows the list's name
and options.
Adds: clicking the modify button in the dialog now properly modifies the
list's name and options in the database.
Adds: TAG string for Log messages.
Bugfix: clicking on the delete list button in the dialog now searches
the list name in the cursor rather than getting it from the dialog view.